+++ src/main/java/gabim/permissionTrackingApplication/dto/Employee/EmployeeListDto.java
@@ -15,7 +15,7 @@ public class EmployeeListDto {
 
     private int id;
 
-    private int tcNo;
+    private long tcNo;
 
     private String name;
 
@@ -27,7 +27,7 @@ public class EmployeeListDto {
 
     private Date birthday;
 
-    private int phoneNumber;
+    private long phoneNumber;
 
     private String email;
