+++ src/main/java/gabim/permissionTrackingApplication/service/EmployeeService.java
@@ -19,7 +19,7 @@ public interface EmployeeService {
 
 	DataResult<EmployeeDto> getById(Integer id);
 
-	DataResult<EmployeeDto> getByTcNo(Integer tcNo);
+	DataResult<EmployeeDto> getByTcNo(Long tcNo);
